more work on kernelization

+++ src/main/java/splitstree6/xtra/kernelize/Kernelize.java
@@ -19,18 +19,17 @@
 
 package splitstree6.xtra.kernelize;
 
+import jloda.graph.Edge;
 import jloda.graph.Graph;
 import jloda.graph.Node;
 import jloda.graph.NodeArray;
+import jloda.graph.algorithms.IsDAG;
+import jloda.phylo.LSAUtils;
 import jloda.phylo.NewickIO;
 import jloda.phylo.PhyloTree;
 import jloda.phylo.algorithms.ClusterPoppingAlgorithm;
-import jloda.util.Basic;
-import jloda.util.BitSetUtils;
-import jloda.util.IteratorUtils;
-import jloda.util.StringUtils;
+import jloda.util.*;
 import jloda.util.progress.ProgressListener;
-import splitstree6.autumn.Cluster;
 import splitstree6.data.TaxaBlock;
 import splitstree6.splits.TreesUtils;
 
@@ -58,7 +57,7 @@ public class Kernelize {
 										int maxNumberOfResults, boolean mutualRefinement) throws IOException {
 
 		// setup incompatibility graph
-		var incompatibilityGraph = computeClusterIncompatibilityGraph(inputTrees);
+		var incompatibilityGraph = ClusterIncompatibilityGraph.apply(inputTrees);
 		System.err.println("Compatible clusters:   " + incompatibilityGraph.nodeStream().filter(v -> v.getDegree() == 0).count());
 		System.err.println("Incompatible clusters: " + incompatibilityGraph.nodeStream().filter(v -> v.getDegree() > 0).count());
 
@@ -94,6 +93,9 @@ public class Kernelize {
 			for (var component : components) {
 				if (component.getNumberOfNodes() > 1) {
 					System.err.println("Component " + (++count) + ": " + component.getNumberOfNodes() + " clusters");
+					for (var v : component.nodes())
+						System.err.print(" " + v.getInfo());
+					System.err.println();
 					var reducedTrees = extractTrees(component);
 					System.err.println(reducedTrees.report());
 				}
@@ -101,44 +103,83 @@ public class Kernelize {
 		}
 
 		System.err.println("Blob tree: " + NewickIO.toString(blobTree, false) + ";");
+		if (true) {
+			var tree = new PhyloTree(blobTree);
+			for (var v : tree.nodes()) {
+				if (tree.hasTaxa(v)) {
+					tree.setLabel(v, "t" + tree.getTaxon(v));
+				}
+			}
+			System.err.println("Blob tree: " + NewickIO.toString(tree, false) + ";");
+			checkNetwork("Blob tree", blobTree);
+		}
 
 		// run the algorithm on all components:
 		try (NodeArray<TreesAndTaxonClasses> blobNetworksMap = blobTree.newNodeArray()) {
 			for (var component : components) {
 				if (component.getNumberOfNodes() > 1) {
-					var reducedTrees = extractTrees(component);
+					var reducedTreesAndTaxonClasses = extractTrees(component);
 					if (mutualRefinement) {
-						mutualRefinement(reducedTrees.trees());
+						var result = MutualRefinement.apply(reducedTreesAndTaxonClasses.trees(), true);
+						reducedTreesAndTaxonClasses.trees().clear();
+						reducedTreesAndTaxonClasses.trees().addAll(result);
 					}
 
-					System.err.println("Input trees: " + NewickIO.toString(reducedTrees.trees(), false));
-					var networks = algorithm.apply(reducedTrees.trees(), progress);
-					System.err.println("Output networks: " + NewickIO.toString(networks, false));
+					System.err.println("Subproblem input: " + NewickIO.toString(reducedTreesAndTaxonClasses.trees(), false));
+					for (var tree : reducedTreesAndTaxonClasses.trees()) {
+						checkNetwork("subproblem input", tree);
+					}
 
+					var networks = algorithm.apply(reducedTreesAndTaxonClasses.trees(), progress);
+					// restore equivalent taxa
+					if (false) {
+						var taxonEquivalenceClassMap = new HashMap<Integer, BitSet>();
+						for (var equivalenceClass : reducedTreesAndTaxonClasses.taxonClasses()) {
+							if (equivalenceClass.cardinality() > 1) {
+								for (var t : BitSetUtils.members(equivalenceClass)) {
+									taxonEquivalenceClassMap.put(t, equivalenceClass);
+								}
+							}
+						}
+						for (var network : networks) {
+							var leaves = IteratorUtils.asList(network.leaves());
+							for (var v : leaves) {
+								var taxonId = network.getTaxon(v);
+								if (taxonEquivalenceClassMap.containsKey(taxonId)) {
+									var p = v.getParent();
+									for (var t : BitSetUtils.members(taxonEquivalenceClassMap.get(taxonId))) {
+										if (t != taxonId) {
+											var w = network.newNode();
+											w.setLabel("t" + t);
+											network.addTaxon(w, t);
+											network.newEdge(p, w);
+										}
+									}
+								}
+							}
+						}
+					}
 					if (true) { // todo: algorithm doesn't return taxon ids
 
 						if (IteratorUtils.size(networks.iterator().next().getTaxa()) == 0) {
-							var labelTaxonMap = new HashMap<String, Integer>();
-							for (var tree : reducedTrees.trees()) {
-								for (var v : tree.nodes()) {
-									var label = tree.getLabel(v);
-									if (label != null)
-										labelTaxonMap.put(label, tree.getTaxon(v));
-								}
-							}
 							for (var network : networks) {
 								for (var v : network.nodes()) {
 									var label = network.getLabel(v);
-									if (label != null) {
-										var taxonId = labelTaxonMap.get(label);
+									if (label != null && NumberUtils.isInteger(label.substring(1))) {
+										var taxonId = NumberUtils.parseInt(label.substring(1));
 										network.addTaxon(v, taxonId);
 									}
 								}
 							}
 						}
 					}
-					var subnetworks = new TreesAndTaxonClasses(networks, reducedTrees.taxonClasses());
-					var donorTaxa = BitSetUtils.union(reducedTrees.taxonClasses());
+					System.err.println("Subproblem networks: " + NewickIO.toString(networks, false));
+					for (var network : networks) {
+						checkNetwork("Subproblem networks", network);
+					}
+
+					var subnetworks = new TreesAndTaxonClasses(networks, reducedTreesAndTaxonClasses.taxonClasses());
+					var donorTaxa = BitSetUtils.union(reducedTreesAndTaxonClasses.taxonClasses());
 					var acceptorNode = clusterNodeMap.get(donorTaxa);
 					blobNetworksMap.put(acceptorNode, subnetworks);
 				}
@@ -158,6 +199,11 @@ public class Kernelize {
 				for (var e : network.edges()) {
 					network.setReticulate(e, e.getTarget().getInDegree() > 1);
 				}
+				LSAUtils.computeLSAChildrenMap(network, network.newNodeArray());
+				System.err.println(NewickIO.toString(network, false) + ";");
+				if (!checkNetwork("output", network))
+					throw new IOException("Bad output");
+
 			}
 			return networks;
 		}
@@ -186,6 +232,7 @@ public class Kernelize {
 
 					var blobChildren = IteratorUtils.asList(blobNode.children());
 					var addedNodes = new ArrayList<Node>();
+					var addedEdges = new ArrayList<Edge>();
 
 					try (NodeArray<Node> donor2acceptorMap = donorNetwork.newNodeArray()) {
 						for (var v : donorNetwork.nodes()) {
@@ -194,7 +241,7 @@ public class Kernelize {
 							addedNodes.add(w);
 						}
 						for (var e : donorNetwork.edges()) {
-							blobTree.newEdge(donor2acceptorMap.get(e.getSource()), donor2acceptorMap.get(e.getTarget()));
+							addedEdges.add(blobTree.newEdge(donor2acceptorMap.get(e.getSource()), donor2acceptorMap.get(e.getTarget())));
 						}
 						for (var e : IteratorUtils.asList(blobNode.outEdges())) {
 							blobTree.deleteEdge(e);
@@ -203,24 +250,53 @@ public class Kernelize {
 						blobTree.newEdge(blobNode, donor2acceptorMap.get(donorRoot));
 						for (var v : donorNetwork.nodeStream().filter(Node::isLeaf).toList()) {
 							var cluster = BitSetUtils.asBitSet(donorNetwork.getTaxa(v));
-							for (var set : taxonClasses) {
-								if (cluster.intersects(set)) {
-									cluster.or(set);
-									break;
+							if (true) {
+								for (var set : taxonClasses) {
+									if (cluster.intersects(set)) {
+										cluster.or(set);
+										break;
+									}
 								}
 							}
-							var target = clusterNodeMap.get(cluster);
-							if (target == null)
-								System.err.println("null for: " + StringUtils.toString(cluster));
-							blobTree.newEdge(donor2acceptorMap.get(v), target);
+							var targets = new HashSet<Node>();
+							if (clusterNodeMap.containsKey(cluster))
+								targets.add(clusterNodeMap.get(cluster));
+							else {
+								for (var t : BitSetUtils.members(cluster)) {
+									targets.add(clusterNodeMap.get(BitSetUtils.asBitSet(t)));
+								}
+							}
+							if (targets.size() == 1) {
+								addedEdges.add(blobTree.newEdge(donor2acceptorMap.get(v), targets.iterator().next()));
+							} else {
+								for (var target : targets) {
+									addedEdges.add(blobTree.newEdge(donor2acceptorMap.get(v), target));
+								}
+							}
 						}
 					}
 					insertRec(blobTree, clusterNodeMap, blobNetworksMap, numberOfBlobs, resolvedBlobs, maxNumberOfResults, networks);
-					if (resolvedBlobs.size() == numberOfBlobs) {
-						networks.add(new PhyloTree(blobTree));
+					if (!checkNetwork("blob tree", blobTree)) {
+						System.err.println("What?");
+						var copy = new PhyloTree(blobTree);
+						for (var v : copy.nodes()) {
+							if (v.isLeaf() && copy.getLabel(v) == null)
+								copy.setLabel(v, "What?");
+							if (v.getInDegree() > 1) {
+								for (var e : v.inEdges())
+									copy.setReticulate(e, true);
+							}
+						}
+						System.err.println(NewickIO.toString(copy, false) + ";");
+					}
+					networks.add(new PhyloTree(blobTree));
+					for (var e : addedEdges) {
+						if (e.getOwner() != null)
+							blobTree.deleteEdge(e);
 					}
 					for (var v : addedNodes) {
-						blobTree.deleteNode(v);
+						if (v.getOwner() != null)
+							blobTree.deleteNode(v);
 					}
 					for (var v : blobChildren) {
 						blobTree.newEdge(blobNode, v);
@@ -230,48 +306,6 @@ public class Kernelize {
 				return;
 			}
 		}
-	}
-
-	/**
-	 * computes the cluster incompatibility graph for a collection of trees
-	 * in graph, v.getInfo() contains cluster as bit set
-	 * in graph, v.getData() contains indices of all trees that have the cluster, as bit set
-	 *
-	 * @param trees input trees
-	 * @return graph
-	 */
-	public static Graph computeClusterIncompatibilityGraph(Collection<PhyloTree> trees) {
-		var graph = new Graph();
-		// v.getInfo() contains cluster as bit set
-		// v.getData() contains indices of all trees that have the cluster, as bit set
-		var clusterNodeMap = new HashMap<BitSet, Node>();
-		// setup nodes:
-		var treeId = 0;
-		for (var tree : trees) {
-			treeId++;
-			try (var clusterMap = TreesUtils.extractClusters(tree)) {
-				for (var cluster : clusterMap.values()) {
-					var v = clusterNodeMap.computeIfAbsent(cluster, graph::newNode);
-					var which = (BitSet) v.getData();
-					if (which == null) {
-						which = new BitSet();
-						v.setData(which);
-					}
-					which.set(treeId);
-				}
-			}
-		}
-		// setup edges:
-		for (var v = graph.getFirstNode(); v != null; v = v.getNext()) {
-			var cv = (BitSet) v.getInfo();
-			for (var w = v.getNext(); w != null; w = w.getNext()) {
-				var cw = (BitSet) w.getInfo();
-				if (Cluster.incompatible(cv, cw)) {
-					graph.newEdge(v, w);
-				}
-			}
-		}
-		return graph;
 	}
 
 	/**
@@ -327,18 +361,18 @@ public class Kernelize {
 	/**
 	 * extracts the list of distinct trees from a connected cluster-incompatibility graph
 	 *
-	 * @param graph component
+	 * @param incompatibilityComponent component
 	 * @return distinct trees
 	 */
-	private static TreesAndTaxonClasses extractTrees(Graph graph) {
+	private static TreesAndTaxonClasses extractTrees(Graph incompatibilityComponent) {
 		var treeIds = new BitSet();
-		for (var v : graph.nodes()) {
-			if (v.getData() instanceof BitSet bits) {
+		for (var v : incompatibilityComponent.nodes()) {
+			if (v.getData() instanceof BitSet bits) { //getData(): trees containing cluster assoicated with v
 				treeIds.or(bits);
 			}
 		}
 
-		var allClusters = graph.nodeStream().map(v -> (BitSet) v.getInfo()).collect(Collectors.toSet());
+		var allClusters = incompatibilityComponent.nodeStream().map(v -> (BitSet) v.getInfo()).collect(Collectors.toSet());
 		var allTaxa = BitSetUtils.union(allClusters);
 		var taxonClasses = computeTaxonEquivalenceClasses(allTaxa, allClusters);
 		var reducedTaxa = BitSetUtils.asBitSet(taxonClasses.stream().mapToInt(t -> t.nextSetBit(1)).toArray());
@@ -349,7 +383,7 @@ public class Kernelize {
 		for (var treeId : BitSetUtils.members(treeIds)) {
 			var clusters = new HashSet<BitSet>();
 			var taxa = new BitSet();
-			for (var v : graph.nodes()) {
+			for (var v : incompatibilityComponent.nodes()) {
 				if (v.getData() instanceof BitSet bits && bits.get(treeId) && v.getInfo() instanceof BitSet cluster) {
 					clusters.add(BitSetUtils.copy(cluster));
 					taxa.or(cluster);
@@ -413,36 +447,49 @@ public class Kernelize {
 
 		return new TreesAndTaxonClasses(trees, taxonClasses);
 	}
-
-	/**
-	 * performs mutatual refinement of all input trees
-	 *
-	 * @param trees input trees
-	 * @return mut
-	 */
-	public static void mutualRefinement(Collection<PhyloTree> trees) {
-		var incompatibilityGraph = computeClusterIncompatibilityGraph(trees);
-
-		var compatibleClusters = incompatibilityGraph.nodeStream()
-				.filter(v -> v.getDegree() == 0) // compatible with all
-				.filter(v -> ((BitSet) v.getData()).cardinality() < trees.size()) // not in all trees
-				.map(v -> (BitSet) v.getInfo()).toList();
-
-		if (!compatibleClusters.isEmpty()) {
-			var result = new ArrayList<PhyloTree>();
-			for (var tree : trees) {
-				// todo: insert clusters into existing tree
-				var clusters = TreesUtils.collectAllHardwiredClusters(tree);
-				if (!clusters.containsAll(compatibleClusters)) {
-					clusters.addAll(compatibleClusters);
-					var newTree = new PhyloTree();
-					ClusterPoppingAlgorithm.apply(clusters, newTree);
-					result.add(newTree);
-				} else result.add(tree);
-			}
-			trees.clear();
-			trees.addAll(result);
+	
+	public static boolean checkNetwork(String label, PhyloTree network) {
+		var ok = true;
+		if (!IsDAG.apply(network)) {
+			System.err.println(label + ": Error: Is not a DAG!");
+			ok = false;
 		}
+		if (network.getRoot() == null) {
+			System.err.println(label + ": Error: root not set");
+			ok = false;
+		} else {
+			if (network.getRoot().getInDegree() > 0) {
+				System.err.println(label + ": Error: root has indegree " + network.getRoot().getInDegree());
+				ok = false;
+			}
+		}
+
+		var roots = network.nodeStream().filter(v -> v.getInDegree() == 0).count();
+		if (roots > 1) {
+			System.err.println(label + ": Error: Too many potential roots: " + roots);
+			ok = false;
+		}
+		var isolatedNodes = network.nodeStream().filter(v -> v.getDegree() == 0).count();
+		if (isolatedNodes > 1) {
+			System.err.println(label + ": Error: has isolated nodes: " + isolatedNodes);
+			ok = false;
+		}
+
+		var leavesWithoutTaxa = network.nodeStream().filter(v -> v.getOutDegree() == 0 && !network.hasTaxa(v)).count();
+		if (leavesWithoutTaxa > 0) {
+			System.err.println(label + ": Error: leaves without taxa: " + leavesWithoutTaxa);
+			ok = false;
+		}
+		var internalWithTaxa = network.nodeStream().filter(v -> v.getOutDegree() != 0 && network.hasTaxa(v)).count();
+
+		if (internalWithTaxa > 0) {
+			System.err.println(label + ": Error: internal nodes with taxa: " + internalWithTaxa);
+			ok = false;
+		}
+		if (ok)
+			System.err.println(label + ": looks ok");
+		return ok;
+
 	}
 
 	/**
